Organize core domain modules
